requirements.txt
streamlit>=1.30,<1.51
rdkit-pypi>=2022.9
pandas>=2.0
numpy>=1.24
plotly>=5.18
scikit-learn>=1.3
requests
